src/db/requirements.txt
elasticsearch
spacy
boto3
python-dotenv
datetime